Potential design for cfrl. Includes working examples for MNIST and ADULT for tflow backend.
