+++ alibi/explainers/backends/tflow/cfrl_base.py
@@ -1,20 +1,17 @@
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-from torch.utils.data import DataLoader, Dataset
-
 import numpy as np
-from typing import List, Dict, Callable, Union, Optional, TYPE_CHECKING
+import tensorflow as tf
+import tensorflow.keras as keras
+from typing import Any, List, Dict, Callable, Union, Optional, TYPE_CHECKING
 
 from alibi.explainers.backends.cfrl_base import CounterfactualRLDataset
-from alibi.models.pytorch.actor_critic import Actor, Critic
+from alibi.models.tensorflow.actor_critic import Actor, Critic
 
 if TYPE_CHECKING:
     from alibi.explainers.cfrl_base import NormalActionNoise
 
 
-class PtCounterfactualRLDataset(CounterfactualRLDataset, Dataset):
-    """ Pytorch backend datasets. """
+class TfCounterfactualRLDataset(CounterfactualRLDataset, keras.utils.Sequence):
+    """ Tensorflow backend datasets. """
 
     def __init__(self,
                  x: np.ndarray,
@@ -22,7 +19,8 @@ class PtCounterfactualRLDataset(CounterfactualRLDataset, Dataset):
                  predict_func: Callable,
                  conditional_func: Callable,
                  num_classes: int,
-                 batch_size: int) -> None:
+                 batch_size: int,
+                 shuffle: bool = True) -> None:
         """
         Constructor.
 
@@ -37,13 +35,15 @@ class PtCounterfactualRLDataset(CounterfactualRLDataset, Dataset):
             Prediction function. The classifier function should expect the input in the original format and preprocess
             it internally in the `predict_func` if necessary.
         conditional_func
-            Conditional function generator. Given an preprocessed input array, the functions generates a conditional
+            Conditional function generator. Given an preprocesed input array, the functions generates a conditional
             array.
         num_classes
             Number of classes in the dataset.
         batch_size
             Dimension of the batch used during training. The same batch size is used to infer the classification
             labels of the input dataset.
+        shuffle
+            Whether to shuffle the dataset each epoch. `True` by default.
         """
         super().__init__()
 
@@ -53,60 +53,71 @@ class PtCounterfactualRLDataset(CounterfactualRLDataset, Dataset):
         self.conditional_func = conditional_func
         self.num_classes = num_classes
         self.batch_size = batch_size
+        self.shuffle = shuffle
+        self.indexes = None
 
         # Infer the classification labels of the input dataset. This is performed in batches.
-        self.y_m = PtCounterfactualRLDataset.predict_batches(x=self.x,
+        self.y_m = TfCounterfactualRLDataset.predict_batches(x=self.x,
                                                              predict_func=self.predict_func,
                                                              batch_size=self.batch_size)
 
-        # Preprocess the input data.
+        # Preprocess data.
         self.x = self.preprocessor(self.x)
 
-    def __len__(self):
-        return self.x.shape[0]
+        # Generate shuffled indexes.
+        self.on_epoch_end()
+
+    def on_epoch_end(self) -> None:
+        """
+        This method is called every epoch and performs dataset shuffling.
+        """
+        self.indexes = np.arange(self.x.shape[0])
+
+        if self.shuffle:
+            np.random.shuffle(self.indexes)
+
+    def __len__(self) -> int:
+        return self.x.shape[0] // self.batch_size
 
     def __getitem__(self, idx) -> Dict[str, np.ndarray]:
         self.num_classes = np.clip(self.num_classes, a_min=0, a_max=2)  # TODO: remove this
 
-        # Generate random target class.
-        y_t = np.random.randint(low=0, high=self.num_classes, size=1).item()
-        data = {
-            "x": self.x[idx],
-            "y_m": self.y_m[idx],
+        # Select indices to be returned.
+        indexes = self.indexes[idx * self.batch_size:(idx + 1) * self.batch_size]
+
+        # Generate random target.
+        y_t = np.random.randint(low=0, high=self.num_classes, size=self.batch_size)
+
+        # compute conditional vector.
+        c = self.conditional_func(self.x[idx * self.batch_size: (idx + 1) * self.batch_size])
+
+        return {
+            "x": self.x[indexes],
+            "y_m": self.y_m[indexes],
             "y_t": y_t,
+            "c": c
         }
 
-        # Construct conditional vector.
-        c = self.conditional_func(self.x[idx:idx+1])
-        if c is not None:
-            data.update({"c": c.reshape(-1)})
 
-        return data
-
-
-def get_device() -> torch.device:
-    """
-    Checks if cuda is available. If available, use cuda by default, else use cpu.
-
-    Returns
-    -------
-    Device to be used.
-    """
-    return torch.device("cuda" if torch.cuda.is_available() else "cpu")
-
-
-def get_optimizer(model: nn.Module, lr: float = 1e-3) -> torch.optim.Optimizer:
+def get_optimizer(model: Optional[keras.Model] = None, lr: float = 1e-3) -> keras.optimizers.Optimizer:
     """
     Constructs default Adam optimizer.
+
+    Parameters
+    ----------
+    model
+        Model to get the optimizer for. Not required for `tensorflow` backend.
+    lr
+        Learning rate.
 
     Returns
     -------
     Default optimizer.
     """
-    return torch.optim.Adam(model.parameters(), lr=lr)
+    return keras.optimizers.Adam(learning_rate=lr)
 
 
-def get_actor(hidden_dim: int, output_dim: int, **kwargs) -> nn.Module:
+def get_actor(hidden_dim: int, output_dim: int) -> keras.layers.Layer:
     """
     Constructs the actor network.
 
@@ -124,7 +135,7 @@ def get_actor(hidden_dim: int, output_dim: int, **kwargs) -> nn.Module:
     return Actor(hidden_dim=hidden_dim, output_dim=output_dim)
 
 
-def get_critic(hidden_dim: int):
+def get_critic(hidden_dim: int) -> keras.layers.Layer:
     """
     Constructs the critic network.
 
@@ -140,7 +151,7 @@ def get_critic(hidden_dim: int):
     return Critic(hidden_dim=hidden_dim)
 
 
-def sparsity_loss(x_hat_cf: torch.Tensor, x: torch.Tensor) -> Dict[str, torch.Tensor]:
+def sparsity_loss(x_hat_cf: tf.Tensor, x: tf.Tensor) -> Dict[str, tf.Tensor]:
     """
     Default L1 sparsity loss.
 
@@ -155,10 +166,10 @@ def sparsity_loss(x_hat_cf: torch.Tensor, x: torch.Tensor) -> Dict[str, torch.Te
     -------
     L1 sparsity loss.
     """
-    return {"sparsity_loss": F.l1_loss(x_hat_cf, x)}
+    return {"sparsity_loss": tf.reduce_mean(tf.abs(x_hat_cf - x))}
 
 
-def consistency_loss(z_cf_pred: torch.Tensor, z_cf_tgt: torch.Tensor):
+def consistency_loss(z_cf_pred: tf.Tensor, z_cf_tgt: tf.Tensor):
     """
     Default 0 consistency loss.
 
@@ -182,8 +193,7 @@ def data_generator(x: np.ndarray,
                    conditional_func: Callable,
                    num_classes: int,
                    batch_size: int,
-                   shuffle: bool,
-                   num_workers: int,
+                   shuffle: bool = True,
                    **kwargs):
     """
     Constructs a tensorflow data generator.
@@ -208,18 +218,13 @@ def data_generator(x: np.ndarray,
         labels of the input dataset.
     shuffle
         Whether to shuffle the dataset each epoch. `True` by default.
-    num_workers
-        Number of worker processes to be created.
     """
-    dataset = PtCounterfactualRLDataset(x=x, preprocessor=ae_preprocessor, predict_func=predict_func,
-                                        conditional_func=conditional_func, num_classes=num_classes,
-                                        batch_size=batch_size)
-    return DataLoader(dataset=dataset, batch_size=batch_size, num_workers=num_workers,
-                      shuffle=shuffle, drop_last=True)
+    return TfCounterfactualRLDataset(x=x, preprocessor=ae_preprocessor, predict_func=predict_func,
+                                     conditional_func=conditional_func, num_classes=num_classes,
+                                     batch_size=batch_size, shuffle=shuffle)
 
 
-@torch.no_grad()
-def encode(x: torch.Tensor, ae: nn.Module, device: torch.device, **kwargs):
+def encode(x: Union[tf.Tensor, np.ndarray], ae: keras.Model, **kwargs):
     """
     Encodes the input tensor.
 
@@ -229,19 +234,15 @@ def encode(x: torch.Tensor, ae: nn.Module, device: torch.device, **kwargs):
         Input to be encoded.
     ae
         Pre-trained autoencoder.
-    device
-        Device to send data to.
 
     Returns
     -------
     Input encoding.
     """
-    ae.eval()
-    return ae.encoder(x.float().to(device))
+    return ae.encoder(x, training=False)
 
 
-@torch.no_grad()
-def decode(z: torch.Tensor, ae: nn.Module, device: torch.device, **kwargs):
+def decode(z: Union[tf.Tensor, np.ndarray], ae: keras.Model, **kwargs):
     """
     Decodes an embedding tensor.
 
@@ -251,27 +252,21 @@ def decode(z: torch.Tensor, ae: nn.Module, device: torch.device, **kwargs):
         Embedding tensor to be decoded.
     ae
         Pre-trained autoencoder.
-    device
-        Device to sent data to.
 
     Returns
     -------
     Embedding tensor decoding.
     """
-    ae.eval()
-    return ae.decoder(z.float().to(device))
+    return ae.decoder(z, training=False)
 
 
-@torch.no_grad()
-def generate_cf(z: torch.Tensor,
-                y_m: torch.Tensor,
-                y_t: torch.Tensor,
-                c: Optional[torch.Tensor],
+def generate_cf(z: Union[np.ndarray, tf.Tensor],
+                y_m: Union[np.ndarray, tf.Tensor],
+                y_t: Union[np.ndarray, tf.Tensor],
+                c: Optional[Union[np.ndarray, tf.Tensor]],
                 num_classes: int,
-                ae: nn.Module,
-                actor: nn.Module,
-                device: torch.device,
-                **kwargs) -> torch.Tensor:
+                actor: keras.Model,
+                **kwargs) -> tf.Tensor:
     """
     Generates counterfactual embedding.
 
@@ -287,88 +282,79 @@ def generate_cf(z: torch.Tensor,
         Conditional tensor.
     num_classes
         Number of classes to be considered.
-    ae
-        Pre-trained autoencoder.
     actor
         Actor network. The model generates the counterfactual embedding.
-    device
-        Device object to be used.
 
     Returns
     -------
     z_cf
         Counterfactual embedding.
     """
-    # Set autoencoder and actor to evaluation mode.
-    ae.eval()
-    actor.eval()
-
-    # Transform classification labels into one-hot encoding.
-    y_m_ohe = F.one_hot(y_m.long(), num_classes=num_classes).float().to(device)
-    y_t_ohe = F.one_hot(y_t.long(), num_classes=num_classes).float().to(device)
+    # Transform to one hot encoding model's prediction and the given target
+    y_m_ohe = tf.one_hot(tf.cast(y_m, dtype=tf.int32), depth=num_classes, dtype=tf.float32)
+    y_t_ohe = tf.one_hot(tf.cast(y_t, dtype=tf.int32), depth=num_classes, dtype=tf.float32)
 
     # Concatenate z_mean, y_m_ohe, y_t_ohe to create the input representation for the projection network (actor).
-    state = [z.view(z.shape[0], -1), y_m_ohe, y_t_ohe] + ([c.float().to(device)] if (c is not None) else [])
-    state = torch.cat(state, dim=1)
+    state = [tf.reshape(z, (z.shape[0], -1)), y_m_ohe, y_t_ohe] + \
+            ([tf.constant(c, dtype=tf.float32)] if (c is not None) else [])
+    state = tf.concat(state, axis=1)
 
     # Pass the new input to the projection network (actor) to get the counterfactual embedding
-    z_cf = actor(state)
+    z_cf = actor(state, training=False)
     return z_cf
 
 
-def add_noise(z_cf: torch.Tensor,
+def add_noise(z_cf: Union[tf.Tensor, np.ndarray],
               noise: 'NormalActionNoise',
               act_low: float,
               act_high: float,
               step: int,
               exploration_steps: int,
-              device: torch.device,
-              **kwargs) -> torch.Tensor:
+              **kwargs) -> tf.Tensor:
     """
     Add noise to the counterfactual embedding.
 
     Parameters
     ----------
     z_cf
-       Counterfactual embedding.
+        Counterfactual embedding.
     noise
-       Noise generator object.
+        Noise generator object.
     act_low
-        Action lower bound.
+        Noise lower bound.
     act_high
-        Action upper bound.
+        Noise upper bound.
     step
-       Training step.
+        Training step.
     exploration_steps
-       Number of exploration steps. For the first `exploration_steps`, the noised counterfactul embedding
-       is sampled uniformly at random.
-    device
-        Device to send data to.
+        Number of exploration steps. For the first `exploration_steps`, the noised counterfactul embedding
+        is sampled uniformly at random.
 
     Returns
     -------
     z_cf_tilde
-       Noised counterfactual embedding.
+        Noised counterfactual embedding.
     """
     # Generate noise.
-    eps = torch.tensor(noise(z_cf.shape)).float().to(device)
+    eps = noise(z_cf.shape)
 
     if step > exploration_steps:
         z_cf_tilde = z_cf + eps
-        z_cf_tilde = torch.clamp(z_cf_tilde, min=act_low, max=act_high)
+        z_cf_tilde = tf.clip_by_value(z_cf_tilde, clip_value_min=act_low, clip_value_max=act_high)
     else:
         # for the first exploration_steps, the action is sampled from a uniform distribution between
         # [act_low, act_high] to encourage exploration. After that, the algorithm returns to the normal exploration.
-        z_cf_tilde = (act_low + (act_high - act_low) * torch.rand_like(z_cf)).to(device)
+        z_cf_tilde = tf.random.uniform(z_cf.shape, minval=act_low, maxval=act_high)
 
     return z_cf_tilde
 
 
-def update_actor_critic(ae: nn.Module,
-                        critic: nn.Module,
-                        actor: nn.Module,
-                        optimizer_critic: torch.optim.Optimizer,
-                        optimizer_actor: torch.optim.Optimizer,
+@tf.function()
+def update_actor_critic(ae: keras.Model,
+                        critic: keras.Model,
+                        actor: keras.Model,
+                        optimizer_critic: keras.optimizers.Optimizer,
+                        optimizer_actor: keras.optimizers.Optimizer,
                         sparsity_loss: Callable,
                         consistency_loss: Callable,
                         coeff_sparsity: float,
@@ -382,8 +368,7 @@ def update_actor_critic(ae: nn.Module,
                         y_t: np.ndarray,
                         c: Optional[np.ndarray],
                         r_tilde: np.ndarray,
-                        device: torch.device,
-                        **kwargs):
+                        **kwargs) -> Dict[str, Any]:
     """
     Training step. Updates actor and critic networks including additional losses.
 
@@ -425,91 +410,77 @@ def update_actor_critic(ae: nn.Module,
         Conditional tensor.
     r_tilde
         Noised counterfactual reward.
-    device
-        Torch device object.
 
     Returns
     -------
     Dictionary of losses.
     """
-    # Set autoencoder to evaluation mode.
-    ae.eval()
-
-    # Set actor and critic to training mode.
-    actor.train()
-    critic.train()
-
     # Define dictionary of losses.
     losses: Dict[str, float] = dict()
 
-    # Transform data to tensors and it device
-    x = torch.tensor(x).float().to(device)
-    x_cf = torch.tensor(x_cf).float().to(device)
-    z = torch.tensor(z).float().to(device)
-    z_cf_tilde = torch.tensor(z_cf_tilde).float().to(device)
-    y_m_ohe = F.one_hot(torch.tensor(y_m, dtype=torch.long), num_classes=num_classes).float().to(device)
-    y_t_ohe = F.one_hot(torch.tensor(y_t, dtype=torch.long), num_classes=num_classes).float().to(device)
-    c = torch.tensor(c).float().to(device) if (c is not None) else None
-    r_tilde = torch.tensor(r_tilde).float().to(device)
+    # Transform classification labels into one-hot encoding.
+    y_m_ohe = tf.one_hot(tf.cast(y_m, tf.int32), depth=num_classes, dtype=tf.float32)
+    y_t_ohe = tf.one_hot(tf.cast(y_t, tf.int32), depth=num_classes, dtype=tf.float32)
 
     # Define state by concatenating the input embedding, the classification label, the target label, and optionally
     # the conditional vector if exists.
-    state = [z, y_m_ohe, y_t_ohe] + ([c.float().to(device)] if (c is not None) else [])
-    state = torch.cat(state, dim=1).to(device)
+    state = [z, y_m_ohe, y_t_ohe] + ([c] if c is not None else [])
+    state = tf.concat(state, axis=1)
 
-    # Define input for critic, compute q-values and append critic's loss.
-    input_critic = torch.cat([state, z_cf_tilde], dim=1).float()
-    output_critic = critic(input_critic).squeeze(1)
-    loss_critic = F.mse_loss(output_critic, r_tilde)
-    losses.update({"loss_critic": loss_critic.item()})
+    # Define input for critic and compute q-values.
+    with tf.GradientTape() as tape_critic:
+        input_critic = tf.concat([state, z_cf_tilde], axis=1)
+        output_critic = tf.squeeze(critic(input_critic, training=True), axis=1)
+        loss_critic = tf.reduce_mean(tf.square(output_critic - r_tilde))
+
+    # Append critic's loss.
+    losses.update({"loss_critic": loss_critic})
 
     # Update critic by gradient step.
-    optimizer_critic.zero_grad()
-    loss_critic.backward()
-    optimizer_critic.step()
+    grads_critic = tape_critic.gradient(loss_critic, critic.trainable_weights)
+    optimizer_critic.apply_gradients(zip(grads_critic, critic.trainable_weights))
 
-    # Compute counterfactual embedding.
-    z_cf = actor(state)
+    with tf.GradientTape() as tape_actor:
+        # Compute counterfactual embedding.
+        z_cf = actor(state, training=True)
 
-    # Compute critic's output.
-    critic.eval()
-    input_critic = torch.cat([state, z_cf], dim=1)
-    output_critic = critic(input_critic)
+        # Compute critic's output
+        input_critic = tf.concat([state, z_cf], axis=1)
+        output_critic = critic(input_critic, training=True)
 
-    # Compute actor's loss.
-    loss_actor = -torch.mean(output_critic)
-    losses.update({"loss_actor": loss_actor})
+        # Compute actors' loss.
+        loss_actor = -tf.reduce_mean(output_critic)
+        losses.update({"loss_actor": loss_actor})
 
-    # Decode the output of the actor.
-    x_hat_cf = ae.decoder(z_cf)
+        # Decode the counterfactual embedding.
+        x_hat_cf = ae.decoder(z_cf, training=False)
 
-    # Compute sparsity losses.
-    loss_sparsity = sparsity_loss(x_hat_cf, x)
-    losses.update(loss_sparsity)
+        # Compute sparsity losses and append sparsity loss.
+        loss_sparsity = sparsity_loss(x_hat_cf, x)
+        losses.update(loss_sparsity)
 
-    # Add sparsity loss to the overall actor's loss.
-    for key in loss_sparsity.keys():
-        loss_actor += coeff_sparsity * loss_sparsity[key]
+        # Add sparsity loss to the overall actor loss.
+        for key in loss_sparsity.keys():
+            loss_actor += coeff_sparsity * loss_sparsity[key]
 
-    # Compute consistency loss.
-    z_cf_tgt = encode(x=x_cf, ae=ae, device=device)
-    loss_consistency = consistency_loss(z_cf_pred=z_cf, z_cf_tgt=z_cf_tgt)
-    losses.update(loss_consistency)
+        # Compute consistency loss and append consistency loss.
+        z_cf_tgt = ae.encoder(x_cf, training=False)
+        loss_consistency = consistency_loss(z_cf_pred=z_cf, z_cf_tgt=z_cf_tgt)
+        losses.update(loss_consistency)
 
-    # Add consistency loss to the overall actor loss.
-    for key in loss_consistency.keys():
-        loss_actor += coeff_consistency * loss_consistency[key]
+        # Add consistency loss to the overall actor loss.
+        for key in loss_consistency.keys():
+            loss_actor += coeff_consistency * loss_consistency[key]
 
     # Update by gradient descent.
-    optimizer_actor.zero_grad()
-    loss_actor.backward()
-    optimizer_actor.step()
+    grads_actor = tape_actor.gradient(loss_actor, actor.trainable_weights)
+    optimizer_actor.apply_gradients(zip(grads_actor, actor.trainable_weights))
 
-    # Return dictionary of losses for potential logging.
+    # Return dictionary of losses for potential logging
     return losses
 
 
-def to_numpy(x: Optional[Union[List, np.ndarray, torch.Tensor]]) -> Optional[Union[List[np.ndarray], np.ndarray]]:
+def to_numpy(x: Optional[Union[List, np.ndarray, tf.Tensor]]) -> Optional[Union[List[np.ndarray], np.ndarray]]:
     """
     Converts given tensor to numpy array.
 
@@ -526,29 +497,28 @@ def to_numpy(x: Optional[Union[List, np.ndarray, torch.Tensor]]) -> Optional[Uni
         if isinstance(x, np.ndarray):
             return x
 
-        if isinstance(x, torch.Tensor):
-            return x.detach().cpu().numpy()
+        if isinstance(x, tf.Tensor):
+            return x.numpy()
 
         if isinstance(x, list):
             return [to_numpy(e) for e in x]
 
         return np.array(x)
-
     return None
 
 
-def to_tensor(x: Union[np.ndarray, torch.Tensor], device: torch.device, **kwargs) -> Optional[torch.Tensor]:
+def to_tensor(x: Union[np.array, tf.Tensor], **kwargs) -> Optional[tf.Tensor]:
     """
-    Converts tensor to torch.Tensor
+    Converts tensor to tf.Tensor
 
     Returns
     -------
-    torch.Tensor conversion.
+    tf.Tensor conversion.
     """
     if x is not None:
-        if isinstance(x, torch.Tensor):
-            return x.to(device)
+        if isinstance(x, tf.Tensor):
+            return x
 
-        return torch.tensor(x).to(device)
+        return tf.constant(x)
 
     return None
